components: Move utils outside components
The utils module is not a component.

https://phabricator.endlessm.com/T32136
